pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[project]
name = "app"
version = "0.0.1"

description = "FastApi template"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi==0.111.*",
    "uvicorn==0.29.*",
    "sqlalchemy==2.*",
    "alembic==1.13.*",
    "psycopg==3.*",
    "dishka==1.*",
    "passlib==1.*",
    "pydantic-settings==2.*",
    "bcrypt==4.*",
    "dataclass-factory==2.*",
]
[project.optional-dependencies]
test = [
    "pytest"
]
lint = [
    "ruff",
    "black",
]